fix: queryStringify escape
