+++ src/common-components/utils/helpers/HTTPTransport.ts
@@ -1,3 +1,5 @@
+import { queryStringify } from "./queryStringify";
+
 interface RequestProps {
     url: string;
     options: RequestOptions;
@@ -19,18 +21,6 @@ enum METHODS {
     PUT = 'PUT',
     DELETE = 'DELETE',
 }
-
-const queryStringify = (data: Record<string, string | number>) => {
-    let string = '?';
-    const entries = Object.entries(data);
-    entries.forEach(([key, value], index) => {
-        string += `${key}=${value}`;
-        if (index + 1 !== entries.length) {
-            string += '&';
-        }
-    });
-    return string;
-};
 
 export class HTTPTransport {
     static get = ({ url, options = {} }: RequestProps) => {
